Adjust PlayAtTimeStamp
Fixed incompatibility with the TogglePause method with a work around to detect the sound being nearly finished. This has to be done because Untiy provides no built in solution for a AudioClip to be finished, besides isPlaying which also becomes false when pausing the sound.

Additionaly made default method params with constants.