Make AudioManager into a ServiceLocator
Added classes and methods, to make the AudioManager as well as it's dependencies a Service Locator. Meaning a global static class provides access to the AudioManager where a decorator instance that wraps the AudioManager is now used for logging as well as a NullAudioManager that throws a specific error message on metod calls if the user forgot to attach the AudioManagerSettings script to a gameObject in the scene.