Add ILogger interface
Added ILogger interface so you can theoritcally replace the LoggedAudioManager with your own implementation of ILogger that for example logs into a file. Additonaly updated MethodCall.cs to use string.Join as well and replaced Action for callbacks with a custom delegate to make calling a non static method possible.